fix: move behavior tests to tests/behavior/, detect missing contractType on JS postconditions
1. Move pipeline and fnmatch behavior tests to tests/behavior/ per
   CLAUDE.md convention
2. Add runtime guard: if a contract has tool + check with 2+ params
   but no contractType: "post", throw EdictumConfigError instead of
   silently classifying as precondition. Protects JS consumers who
   skip TypeScript's compile-time enforcement.

244 tests passing.
